Refactor episode ID handling and cleanup
Removed `[Serializable]` and `ISerializable` from `StoredId` in
`InvalidEpisodeIdFormatException.cs`, simplifying the struct.

Updated `_isNotEmpty` logic in `SeasonEpisodeId` to depend only on
`season > 0` and adjusted `IsValid` to include `EpisodeRange.IsValid()`.

Refactored `EpisodeRange`:
- Replaced private fields with auto-implemented properties.
- Added XML documentation for clarity.
- Simplified constructor and validation logic.
- Improved `ToString` formatting.

Removed redundant `EpisodeRange2` struct.

Cleaned up `using` directives and added `[StructLayout(LayoutKind.Auto)]`
to `EpisodeRange` for memory optimization.

Improved code style by replacing magic strings and using `nameof` for
member access.
